Create 7-print_last_char.c

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -10,8 +10,8 @@
 int print_last_digit(int j)
 {
 	int last_char = (j < 0) ? (j * -1) % 10 : j % 10;
-
-	_putchar(last_char + '0');
+	
+	_putchar((int)last_char + '0');
 
 	return (last_char);
 }
